resolved most of Manage state space exhaustion #31
